restructured repo and redesigned database to conform to medallion architecture
