+++ .env
@@ -1,0 +1,9 @@
+AWS_S3_BUCKET_NAME=test-monzo-etl-bucket
+AWS_S3_DATABASE_NAME=test_monzo_dashboard.db
+AWS_S3_LOG_PREFIX=logs
+DB_KEY=test_monzo_dashboard.db
+# LOCAL_DB_PATH=./src/temp/test_monzo_dashboard_local.db
+LOCAL_DB_PATH=/tmp/test_monzo_dashboard_local.db
+LOCAL_DB_NAME=test_monzo_dashboard_local.db
+# LOCAL_LOG_PATH=./src/temp/logs/app.log
+LOCAL_LOG_PATH=/tmp/logs/monzo_etl_test.log
