+++ .gitignore
@@ -4,3 +4,4 @@ data/database_files/
 src/logs/monzo_etl.log
 src/.env
 deploy.bat
+deploy_test.bat
