--- src/api/__init__.py
@@ -1,3 +1,0 @@
-from .monzo_api_client import MonzoAPIClient
-
-__all__ = ['MonzoAPIClient']
