+++ src/extract/__init__.py
@@ -1,0 +1,3 @@
+from .extract import MonzoDataExtractor
+
+__all__ = ['MonzoDataExtractor']
