+++ src/load/__init__.py
@@ -1,0 +1,3 @@
+from .load import MonzoBronzeDataLoader
+
+__all__ = ['MonzoBronzeDataLoader']
