--- src/logs/etl_log.log
@@ -1,16 +1,0 @@
-[SAMPLE DATA]
-2024-12-01 15:14:54,629 - MonzoETL - INFO - Enabled S3 log uploading to bucket: monzo-db-bucket
-2024-12-01 15:14:54,629 - MonzoETL - INFO - Initialising MonzoSQLiteETL with database at src/monzo_dashboard.db
-2024-12-01 15:14:54,629 - MonzoETL - INFO - Starting ETL process. Getting transactions for the previous 30 days
-2024-12-01 15:14:55,162 - MonzoETL - INFO - Successfully downloaded database file from S3
-2024-12-01 15:14:55,163 - MonzoETL - INFO - Initialising database tables
-2024-12-01 15:14:55,164 - MonzoETL - INFO - Database tables created successfully
-2024-12-01 15:14:55,164 - MonzoETL - INFO - Retrieving data from Monzo API
-2024-12-01 15:14:55,721 - MonzoETL - INFO - Processing X transactions
-2024-12-01 15:14:55,733 - MonzoETL - INFO - Processing balance data
-2024-12-01 15:14:55,742 - MonzoETL - INFO - Processing X pots
-2024-12-01 15:14:55,796 - MonzoETL - INFO - Successfully inserted X pots
-2024-12-01 15:14:55,982 - MonzoETL - INFO - Successfully uploaded database file from S3
-2024-12-01 15:14:55,983 - MonzoETL - INFO - ETL process completed successfully
-2024-12-01 15:14:56,075 - MonzoETL - INFO - Forced upload of logs to S3 completed
-2024-12-01 15:22:10,139 - MonzoETL - INFO - Enabled S3 log uploading to bucket: monzo-db-bucket
