+++ src/sql/create_gold_layer.sql
@@ -1,0 +1,6 @@
+CREATE TABLE IF NOT EXISTS gold_monthly_spending (
+    month INTEGER,
+    year INTEGER,
+    total_spend REAL,
+    avg_spend REAL
+);
